backend/pyproject.toml
[project]
name = "day-one"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.121.3",
    "schemathesis>=4.6",
    "email-validator>=2.3",
    "psycopg[binary,pool]>=3.2.0",
    "locust>=2.42",
    "pydantic-settings>=2.0.0",
    "pwdlib[bcrypt]",
    "PyJWT",
    "apscheduler>=3.10.0",
    "python-multipart>=0.0.9",
]
